requirements.txt
Flask==0.12.2
gunicorn==19.6.0
line-bot-sdk==1.3.0
imgurpython==1.1.7 
apiai==1.2.3